Insignificant changes.
git-svn-id: https://openmalaria.googlecode.com/svn/trunk@573 aa1ca794-fd11-11dd-ac65-f74f249cfaac

+++ model/Clinical/ClinicalModel.cpp
@@ -47,13 +47,13 @@ void ClinicalModel::init () {
 }
 
 void ClinicalModel::staticRead (istream& in) {
-  for (int i = 0; i < Global::intervalsPerYear; ++i) {
+  for (size_t i = 0; i < Global::intervalsPerYear; ++i) {
     in >> infantDeaths[i];
     in >> infantIntervalsAtRisk[i];
   }
 }
 void ClinicalModel::staticWrite (ostream& out) {
-  for (int i = 0; i < Global::intervalsPerYear; ++i) {
+  for (size_t i = 0; i < Global::intervalsPerYear; ++i) {
     out << infantDeaths[i] << endl;;
     out << infantIntervalsAtRisk[i] << endl;
   }
